+++ app/commands/disclaimers.py
@@ -1,3 +1,5 @@
+from telegram.ext import ConversationHandler
+
 from app.decorators import register_update, chat_language
 
 
@@ -13,3 +15,6 @@ def disclaimers_command(bot, update, chat_info, _):
                'rates displayed. You should confirm current rates before making '
                'any transactions that could be affected by changes in '
                'the exchange rates.'))
+
+    return ConversationHandler.END
+